suggestion for handling gamertag replacement
The original name_fixes.py was not properly replacing some names for me. For example:
The 64 player `ld` also goes by `fox`.
Adding this to `name_fixes.csv` resulted in a player `bakersfield` being renamed `bakersfox`.

I've changed the file to use a dictionary so that a player's name changes exactly when their full gamertag matches a key in the dictionary.
I also wrote it to check that the Smash-game that the replacement gamertag is associated with has to match the Smash-game of the gamertag being replaced.